refactor: client pages
