+++ src/locales/zh-Hans/messages.po
@@ -734,4 +734,3 @@ msgstr "Mint 按钮还未开放"
 
 msgid "<0/> quoted your <1><2/></1>"
 msgstr "<0/> 引用了你的 <1><2/></1>"
-
